requirements.txt
requests
sentence-transformers
faiss-cpu
fastapi
uvicorn
transformers
numpy
